Refactor: TimePage TimeLabelContainer module.css seperate
