+++ src/styles/create/TimePage.module.css
@@ -1,0 +1,20 @@
+.container {
+  /* width: 92.6%; */
+  height: 100%;
+  margin: 0 8%;
+  padding: 7.4% 0 5.8% 0;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  /* gap: 25px; */
+  justify-content: space-around;
+}
+.imageWrapper {
+  height: 14.4%;
+  max-height: 150px;
+  margin: 0 14%;
+}
+.top_label {
+  font-weight: 1000;
+  font-size: 20px;
+}
